pyproject.toml
[project]
name = "centrifugo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.124.2",
    "uvicorn>=0.38.0",
    "httpx>=0.27.0",
    "aiocent>=4.1.2",
    "websockets>=15.0.1",
    "uvloop>=0.22.1",
    "pyjwt>=2.8.0",
    "granian>=1.7.0",
    "numpy>=1.26.0",
]
